Removes test classes for now-unused Form base classes. Combines tests into remaining base class.
